requirements.txt
amqp==2.2.2
backcall==0.1.0
billiard==3.5.0.3
bleach==2.1.3
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
Django==2.0.4
django-celery-results==1.0.1
django-crispy-forms==1.7.2
django-extensions==2.0.7
django-filter==1.1.0
django-tables2==1.21.2
djangorestframework==3.8.2
entrypoints==0.2.3
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.3.1
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kombu==4.1.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.1
oauthlib==2.0.7
pandocfilters==1.4.2
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PySocks==1.6.8
python-dateutil==2.7.2
pytz==2018.4
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
requests-oauthlib==0.8.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
tweepy==3.6.0
urllib3==1.22
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.2.1
